Suggest s_tau

Suggest a value of the s_tau hyperparameter to roughly target a specified minimum amount of borrowing in the hierarchical model. Only use if a diffuse prior on tau is not feasible.

hbl_s_tau(precision_ratio = 0.5, sigma = 1, n = 100)

Arguments

precision_ratio

Positive numeric vector of elements between 0 and 1 with target precision ratios.

sigma

Positive numeric vector of residual standard deviations.

n

Number of non-missing patients.

Value

Numeric of length equal to length(precision_ratio) and length(sigma), suggested values of s_tau for each element of precision_ratio and sigma.

Details

The target minimum amount of borrowing is expressed in the precision_ratio argument. The precision ratio is a metric that quantifies the amount of borrowing in the hierarchical model. See the "Methods" vignette for details.

See also

Examples

hbl_s_tau(precision_ratio = 0.5, sigma = 1, n = 100)
#> [1] 0.2